+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/StorageLocationService.java
@@ -2,7 +2,6 @@ package uk.ac.bbsrc.tgac.miso.service;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
 
 import uk.ac.bbsrc.tgac.miso.core.data.impl.StorageLocation;
 
@@ -15,8 +14,6 @@ public interface StorageLocationService {
   public List<StorageLocation> listRooms();
 
   public List<StorageLocation> listFreezers();
-
-  public Map<String, Integer> getColumnSizes() throws IOException;
 
   public long createRoom(StorageLocation room) throws IOException;